供应链采购订单

+++ cbs-web/src/main/java/com/jaagro/cbs/web/vo/product/DrugStockVo.java
@@ -42,7 +42,7 @@ public class DrugStockVo implements Serializable {
     /**
      * 库存数量
      */
-    private BigDecimal stockQuantity;
+    private Integer stockQuantity;
 
     /**
      * SKU编码
